Redesign gateway class requirements and structure to support DMO superclasses
A single gateway can now be capable of handling multiple DMO types if
their superclass falls under the recognized data type of the gateway

+++ src/data/gateway/UserGateway.java
@@ -4,12 +4,16 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
-import data.gateway.interfaces.IPersonGateway;
+import data.gateway.interfaces.IUserGateway;
+import data.keys.FriendKey;
 import data.keys.Key;
 import data.keys.PersonKey;
+import domain.model.DomainModelObject;
+import domain.model.Friend;
 import domain.model.Person;
+import domain.model.User;
 
-public class PersonGateway extends IPersonGateway {
+public class UserGateway extends IUserGateway {
 
     /**
      * Finds a person using a person key
@@ -35,10 +39,41 @@ public class PersonGateway extends IPersonGateway {
     }
 
     /**
+     * Finds a person using a person key
+     * 
+     * @param key
+     * @return a person, if one was found
+     */
+    private Friend find(FriendKey key) {
+        try {
+            String sql = "SELECT * FROM friend_map f JOIN persons p ON f.fid = p.id WHERE p.name = ?";
+            PreparedStatement stmt = getConnection().prepareStatement(sql);
+            stmt.setString(1, key.name);
+            ResultSet result = stmt.executeQuery();
+            result.next();
+
+            Friend friend = new Friend(result.getString("p.name"),
+                    result.getString("p.name"));
+            return friend;
+        } catch (SQLException e) {
+            e.printStackTrace();
+            System.exit(-1);
+        }
+        return null;
+    }
+    
+    /**
      * Updates the data values of an existing person
      */
     @Override
-    public void update(Person object) {
+    public void update(DomainModelObject dmo) {
+        if (!(dmo instanceof Person))
+        {
+            return;
+        }
+        
+        Person object = (Person)dmo;
+        
         try {
             String sql = "UPDATE persons SET name=?,password=? WHERE id=?";
             PreparedStatement stmt = getConnection().prepareStatement(sql);
@@ -57,7 +92,14 @@ public class PersonGateway extends IPersonGateway {
      * Inserst a new person into the database
      */
     @Override
-    public Result<Person> insert(Person object) {
+    public Result<?> insert(DomainModelObject dmo) {
+        if (!(dmo instanceof Person))
+        {
+            return null;
+        }
+        
+        Person object = (Person)dmo;
+        
         try {
             String sql = "INSERT INTO persons (name, password) VALUES (?, ?)";
             PreparedStatement stmt = getConnection().prepareStatement(sql);
@@ -83,7 +125,7 @@ public class PersonGateway extends IPersonGateway {
                     object = generated;
 
                     PersonKey key = new PersonKey(id);
-                    Result<Person> result = new Result<Person>(generated, key);
+                    Result<User> result = new Result<User>(generated, key);
                     return result;
                 } else {
                     throw new SQLException(
@@ -102,7 +144,14 @@ public class PersonGateway extends IPersonGateway {
      * Deletes a person from the database
      */
     @Override
-    public Key<Person> delete(Person object) {
+    public Key<?> delete(DomainModelObject dmo) {
+        if (!(dmo instanceof User))
+        {
+            return null;
+        }
+        
+        User object = (User)dmo;
+        
         try {
             String sql = "DELETE FROM persons WHERE id=?";
             PreparedStatement stmt = getConnection().prepareStatement(sql);
@@ -122,10 +171,18 @@ public class PersonGateway extends IPersonGateway {
      * Delegating method for finding a person dependent on the type of key given
      */
     @Override
-    public Person find(Key<Person> key) {
+    public DomainModelObject find(Key<?> key) {
         if (key instanceof PersonKey) {
             return find((PersonKey) key);
         }
+        if (key instanceof FriendKey) {
+            return find((FriendKey) key);
+        }
         return null;
+    }
+
+    @Override
+    public Class<User> getType() {
+        return User.class;
     }
 }
